feat: completed the fetch user submission hook
